refactor: created ThreadsGrid.js to follow new file pattern
